requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
lightning==2.0.3  
numpy==1.24.3
pandas==2.0.2
xarray
scikit-learn
optuna
einops
tqdm
pyyaml
wandb
matplotlib>=3.7
geopandas==0.13.2
shapely>=2.0
pyproj>=3.6
cartopy>=0.22
pyshtools>=4.12
fiona==1.9.6
rasterio==1.3.6